DevTools: enforce clang-format on git cl upload
- does a dry run of clang-format to see if detects any style violations
(but doesn't actually fix it)
- if the dry run finds a violation:
  - it runs clang-format a second-time
except it will do in-place edits this time. This fails the git cl upload
because the user needs to commit the formatting changes.
  - also runs autofix for eslint curly braces rule because clang-format can
    split an if block single-statement into multiple lines. Then clang-format is
    run again to align the braces properly.
- based on some research on how others use clang-format, there doesn't seem
a way to get clang-format to both 1) report style violations and 2) make in-place
edits in the same run, thus this slightly more complicated workflow.

BUG=none

Review-Url: https://codereview.chromium.org/2476843002
Cr-Original-Commit-Position: refs/heads/master@{#432307}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2967505bfadaea778a5123a37df0c9b90d9ad6b6