Deflake ontimeout-event-override-after-failure by waiting longer for http
This test flakes when the http server does not respond quick enough. I
suspect this happens on bots that are busy running other tests in
parallel. I was able to simulate this flakiness by lowering the XHR
timeout to 13ms.

This patch increases the XHR timeout from 100ms to 180ms which should
give more time for the test http server to respond. This has no effect
on overall test time because the pass condition is behind a 200ms timer.

After this patch lands, we need to monitor the following:
http://test-results.appspot.com/dashboards/flakiness_dashboard.html#tests=http%2Ftests%2Fxmlhttprequest%2Fontimeout-event-override-after-failure.html
If the flakiness disappears, the lines in TestExpectations should be removed.

BUG=464736,665165

Review-Url: https://codereview.chromium.org/2504713003
Cr-Original-Commit-Position: refs/heads/master@{#432409}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 613135a62748f6d2f875f11b9096c368a3873714