Return one U+fffd for longest subpart of incomplete utf-8 character.
This brings the utf-8 decoder in line w/ the Unicode spec
recommendation, and also with the utf-8 decoder in V8. This
latter bit is the motivation for this change, since either this
or the v8 utf-8 decoder gets used for JavaScript source code, and
they should agree with each other.

More details + spec references in the bug.

BUG=chromium:662822

Review-Url: https://codereview.chromium.org/2498653002
Cr-Original-Commit-Position: refs/heads/master@{#432470}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a8c03cf61f42e45e4c4d2362d97cc2c5abbdc1f