Re-enable foreign-fetch-cors layout test.
With recent fixes in testharness.js and testharnessreport.js this test should
no longer be failing. Maybe this test should be added to SlowTests instead
but for now it seems to be fast enough to not get too close to the actual test
timeout.

BUG=640947

Review-Url: https://codereview.chromium.org/2511473002
Cr-Original-Commit-Position: refs/heads/master@{#432629}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bdd208fae216332c37dfc202d77cf04cd8a1fbd6