binding: Returns a reject promise when |this| is not of the type.
Returns a reject promise when the receiver object is not of type
of the interface, instead of throwing a TypeError.

This CL disables the type check on V8 side and does the type check
on Blink side, then, if the receiver object is not of the type,
converts a TypeError to a reject promise and returns it.

BUG=627309

Review-Url: https://codereview.chromium.org/2441593002
Cr-Original-Commit-Position: refs/heads/master@{#432798}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ef6ea8e2cae02459a206885d1eb6ab2f5b62fbfa