Disable overlay scrollbars to hide them on non-Mac.
In https://codereview.chromium.org/2467693002/ I accidentally removed the code
that disabled/enabled scrollbars based on whether they're hidden. This was
because it caused an invalidation cycle on Mac but I had forgotten that the
reason for disabling scrollbars that are invisible is that painting code uses
this as a signal to "hide" the scrollbar.

This patch brings the code back but makes the behavior dependent on a flag in
ScrollbarTheme. This is needed since the signalling is flipped on Mac. On Mac,
painting code makes the scrollbars invisible which then sets the
scrollbarsHidden flag on the ScrollableArea. On Aura, a timer determines that
the scrollbars need to be hidden and disables them which causes an invalidation
and repaint.

BUG=307091
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2509843004
Cr-Original-Commit-Position: refs/heads/master@{#433300}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4337e5dded29ebb7ef9401cb55e4effcf4f73b69