DCHECK that a valid feature policy is available at the time we query if a feature is enabled
Currently we return the default policy in the case that the FP hasn't been
initialized for a frame. However this might result in returning the wrong value
(because the parent's policy hasn't yet been taken into account). We shouldn't
try to query the policy before it has been initialized from the parent, so
DCHECK that this is the case.

Review-Url: https://codereview.chromium.org/2516233002
Cr-Original-Commit-Position: refs/heads/master@{#433710}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fecfdb85ed5fb437a1d55cbb812f6d884189ad61