Make HarfBuzzShaperTest.ResolveCandidateRunsUnicodeVariants more robust
Some unknown changes to Linux bots figured out that this test is not as
robust as it should be, by relying non-mandatory recommendation.

Since it has started failing on Linux bots, it was disabled in [1].
This patch makes it more robust to fonts that do not follow the
OpenType recommendations and re-enable it.

[1] https://codereview.chromium.org/2516273002

BUG=667147

Review-Url: https://codereview.chromium.org/2520833002
Cr-Original-Commit-Position: refs/heads/master@{#433874}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c3d0368b185088d697bad36d0928a7409d1c970a