Post tasks for sensitive input visibility notifications
If password inputs are hidden and shown multiple times in the same task
(as apparently happens when processing a stylesheet), we don't want to
the omnibox warning to flicker in and out. Thus, this CL posts sensitive
input visibility notifications to the end of the task queue, so that the
omnibox warning updates get coalesced.

BUG=664194
TEST=Visit http://http-password.badssl.com and observe that the omnibox
warning does not flicker in and out twice.

Review-Url: https://codereview.chromium.org/2515373003
Cr-Original-Commit-Position: refs/heads/master@{#434340}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7bdbe352cdbd7cc1dec2b71653e60b47688bd9e1