Add an error page for resources blocked via XSS Auditor.
Currently, when the XSS Auditor blocks a page, we render a lovely white
rectangle for the user. This, though calming and serene, is not terribly
informative.

This patch wires XSS Auditor blocks up to the general error page
mechanism, giving users some clue as to what's going on.

BUG=654794

Review-Url: https://codereview.chromium.org/2425663002
Cr-Original-Commit-Position: refs/heads/master@{#434375}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b5a70d6b91eeabba6752f56cf9d79b8eb57fcf8c