service worker: Persist NavigationPreloadState
This writes NPS to storage. NPS is included as a field in
ServiceWorkerRegistration. I was previously considering how to handle writing
to storage before an active worker exists (and hence before a SWRegistration
exists in storage), but the spec discussion is converging on requiring an
active worker so this should be sufficient.

BUG=649558

Review-Url: https://codereview.chromium.org/2521793004
Cr-Original-Commit-Position: refs/heads/master@{#434411}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 92636463938f6dafce67b97a884d2cb11f27a2cb