Ensure active document for hasEditableLevel
This patch makes hasEditableStyle and hasRichlyEditableStyle return false
directly if document is inactive, because nothing in an inactive document
is editable.

The ultimate goal is to ensure that no code path inspects editable styles
in an inactive document, which will be done by subsequent patches.

BUG=667681

Review-Url: https://codereview.chromium.org/2530753002
Cr-Original-Commit-Position: refs/heads/master@{#434414}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9ceffba8b9056e7a125f5a9fd8eb803816315df7