Emit console warning when element.animate() keyframe value fails to parse
This change adds a return value to the CSS parser to state whether a
CSS value was parsed successfully for a given property.

This information is used to allow calls to element.animate() to warn the
developer when provided keyframe values are invalid.

BUG=618157

Review-Url: https://codereview.chromium.org/2524303002
Cr-Original-Commit-Position: refs/heads/master@{#434490}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1ddf77acbf3b11ee0a3922ca5c1df57d769bac83