Filter webrtc::RTCStats by whitelist in surfacing them to Blink.
This ensures that we don't accidentally leak RTCStats dictionaries
that are not ready to be exposed to the web, and makes it so that
exposing a new dictionary requires a Chromium CL (not just WebRTC
repo CLs).

Note that the white list is not applied to members of whitelisted
dictionaries and that all such members are exposed.

BUG=chromium:627816

Review-Url: https://codereview.chromium.org/2490183002
Cr-Original-Commit-Position: refs/heads/master@{#434501}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 20033c03bd6c08920f93544110a263a13e256c2f