Don't log a console warning when touch-action is used.
To avoid log fatigue for the developer only log the "you are prevent
defaulting a passive event listener" if no touch-action is used. If
touch-action is being used then it is clear that the developer is doing
this to interop with other browsers that don't support touch-action.

BUG=667799

Review-Url: https://codereview.chromium.org/2520423004
Cr-Original-Commit-Position: refs/heads/master@{#434536}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6ea5974532e7cfea654c59181f1819e6805bb368