Worker: Provide a way to access parent frame task runners from a worker thread
To use per-frame task runners on a worker thread, this CL provides a way to
access parent frame task runners from a worker thread. To be specific,
WorkerReportingProxy will have a new interface getParentFrameTaskRunners() and
classes on a worker thread will post a task using the interface.

This is a part of the project to migrate to per-frame task scheduler. A general
design doc about the project is available here:
https://docs.google.com/document/d/10It1DFRP7H3gev9hQA-7dtTcvcMFhrxuCt3-k21yjgQ/edit?usp=sharing

BUG=667310

Review-Url: https://codereview.chromium.org/2513413003
Cr-Original-Commit-Position: refs/heads/master@{#434604}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 409972dd9d6a640a5854d74b142bef19e25a36b5