Worker: Deprecate PostTaskToMainExecutionContext
This CL deprecates PostTaskToMainExecutionContext on InProcessWorkerObjectProxy
and ThreadedWorkletObjectProxy in favor of per-frame task runners. This is a
part of the project to migrate to per-frame scheduler. A general design doc
about the project is available here:
https://docs.google.com/document/d/10It1DFRP7H3gev9hQA-7dtTcvcMFhrxuCt3-k21yjgQ/edit?usp=sharing

This CL also moves UseCounter::count and UseCounter::countDeprecation calls from
DedicatedWorkerGlobalScope to ThreadedMessagingProxyBase because (1) per-frame
task runners don't provide an execution context to a posted task like
ExecutionContextTask (PostTaskToMainExecutionContext) and need an alternative
way to get it, and (2) communications between the main thread and worker threads
should be consolidated to MessaginProxy and ObjectProxy.

BONUS: This cleanup makes it easier to share UseCounter code among workers and
worklets.

BUG=376039, 657694, 667310, 667357

Review-Url: https://codereview.chromium.org/2535653002
Cr-Original-Commit-Position: refs/heads/master@{#434607}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6a8002607cc8e991e382adb29c7c5f13ee55da58