Don't assume current root scroller is a LayoutBox.
There was an edge case here where we might get an arbitrary element as the
document.documentElement. In these cases, the element won't get a LayoutObject
as LayoutView will refuse to attach to it, but when we replace it a layout will
create a non-LayoutBox for it and when RootScroller tries to replace it as the
effective root scroller it needs to do some book keeping using the old root
scroller. It incorrectly assumed it must be a LayoutBox.

This patch now guards against this edge case.

BUG=668553

Review-Url: https://codereview.chromium.org/2531343002
Cr-Original-Commit-Position: refs/heads/master@{#434771}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1a413b53d66a15908a7960fe1d7a3dc3c70a8a5e