Pull up tests for WebP and GIF image decoders.
Just a refactor. No change in behavior, no new tests.

Pull up updateRequiredPreviousFrameAfterFirstDecode,
resumePartialDecodingAfterClearFrameBufferCache, which are shared by
GIFImagerDecoderTest and WEBPImageDecoderTest.

Pull up testAlphaBlending from WEBPImageDecoderTest, so that it can be
reused in APNG tests.

Use ImageDecoderTestHelpers::testByteByByteDecode for GIF's
parseAndDecodeByteByByte test. This tests for the same, and in addition
checks if the baseline hashes are equivalent to the incremental hashes.

Review-Url: https://codereview.chromium.org/2520753002
Cr-Original-Commit-Position: refs/heads/master@{#434828}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 40ce4579e49be92da9fc1569e25f539e44480c97