Abort CompositeEditCommand::moveParagraphs if destination is inside moving range
This patch makes CompositeEditCommand::moveParagraphs abort in the above
mentioned case, because there is no way to move the contents to a destination
inside itself.

There might be a deeper cause of passing such invalid parameter combination
to the function, which is not covered by this patch.

BUG=585048
TEST=webkit_unit_tests --gtest_filter=TypingCommandTest.insertLineBreakWithIllFormedHTML

Review-Url: https://codereview.chromium.org/2531263002
Cr-Original-Commit-Position: refs/heads/master@{#434844}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e1ec97972c1e09d8e472aad4cf225d1ffc5dfef